Refactor routes using Route Groups. Fixed globals.css for each root layout
