+++ src/Components/Structural_Components/HeroSlider.tsx
@@ -3,8 +3,8 @@ import styles from './Styles/hero-slider.module.css'
 
 export default function HeroSlider(){
     return (
-        <Link href="/category/variedades">
-        <div className={styles.hero}>
+        <Link href="/category/variedades" className={styles.hero}>
+        <div>
             HERO/SLIDER
         </div>
         </Link>
