+++ src/Components/Structural_Components/Styles/newsletter-form.module.css
@@ -13,5 +13,10 @@
 }
 
 .form input {
-    background-color: aliceblue;
+    background-color: rgb(255, 250, 236);
+    border: solid;
+    border-color: #000;
+    padding: 5pt;
+    font-size: 14pt;
+    font-weight: 300;
 }
