+++ src/Components/Structural_Components/Styles/product-carousel.module.css
@@ -1,12 +1,12 @@
 .carousel {
-    border:solid;
+    border:none;
     background-color: antiquewhite;
     color: black;
     display: flex;
     flex-direction: row;
     justify-content: space-around;
     align-items: center;
-    width: 100vw;
+    width: 90%;
     height: 35vh;
     padding:1rem;
 }
