+++ src/app/(LP)/thetrackingmonks/layout.tsx
@@ -1,4 +1,4 @@
-import './layout.css'
+import './globals.css'
 
 export default function CheckoutLayout({
   children,
