+++ src/app/(LP)/thetrackingmonks/page.tsx
@@ -3,7 +3,7 @@ import styles from './page.module.css'
 export default function Home() {
   return (
     <main className={styles.main}>
-      PDP
+      LP
     </main>
   )
 }
