+++ src/app/(ecommerce)/category/[slug]/layout.tsx
