+++ src/app/(ecommerce)/category/[slug]/page.tsx
