+++ src/app/(ecommerce)/layout.tsx
@@ -1,8 +1,8 @@
 import './globals.css'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
-import Header from '../Components/Structural_Components/Header'
-import Footer from '../Components/Structural_Components/Footer'
+import Header from '../../Components/Structural_Components/Header'
+import Footer from '../../Components/Structural_Components/Footer'
 import Container from '@/Components/Structural_Components/Container'
 
 const inter = Inter({ subsets: ['latin'] })
