+++ src/app/(ecommerce)/orderPlaced/layout.tsx
