+++ src/app/(ecommerce)/page.module.css
@@ -1,9 +1,8 @@
 .main {
+  height: auto;
+  background-color: antiquewhite;
   display: flex;
   flex-direction: column;
-  justify-content: space-between;
   align-items: center;
-  padding: 6rem;
-  min-height: 100vh;
 }
