+++ src/app/(ecommerce)/product/[slug]/layout.tsx
