+++ src/app/(ecommerce)/product/[slug]/page.tsx
@@ -2,8 +2,8 @@
 
 import { useEffect, useState } from 'react'
 import ProductShippingCalculation from '@/Components/Structural_Components/ProductShippingCalculation'
-import ImageGallery from '../../../Components/Structural_Components/ImageGallery'
-import ProductDetails from '../../../Components/Structural_Components/ProductDetails'
+import ImageGallery from '../../../../Components/Structural_Components/ImageGallery'
+import ProductDetails from '../../../../Components/Structural_Components/ProductDetails'
 import styles from './page.module.css'
 import RelatedProducts from '@/Components/Structural_Components/RelatedProducts'
