--- src/app/page.module.css
@@ -1,5 +1,0 @@
-.main {
-  height: auto;
-  background-color: antiquewhite;
-}
-
